+++ CHANGELOG.rst
@@ -10,6 +10,12 @@ CHANGELOG
 - Fixed fixture 'guillotina'
   [masipcat]
 
+- Make sure that guillotina uses uvloop on starting if its installed
+  [bloodbare]
+
+- Make sure uvicorn uses the same loop as guillotina startup
+  [bloodbare]
+
 - Fix tests in 'test_cache_txn.py' and 'test_setup.py' being skipped
   [masipcat]